% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anitools.R
\name{ANI.dendrogram}
\alias{ANI.dendrogram}
\title{Construct a dendrogram using pairwise ANI values, and identity medoid genomes
using dendrogram coalescence at an ANI threshold (if supplied by the user)}
\usage{
ANI.dendrogram(bactaxRObject, ANI_threshold = 95, label_size = 1e-10,
  color_palette = magma(1000, begin = 0.2), xline = NULL,
  xlinecol = "#20A387FF", xlinetype = "dashed")
}
\arguments{
\item{bactaxRObject}{A bactaxRObject}

\item{ANI_threshold}{ANI threshold at which to assign genomes to
clusters and identify medoid genomes. Defaults to 95.}

\item{label_size}{Size of dendrogram tip labels. Defaults to an
arbitrarily small number (i.e., 1e-10), so that tip labels are hidden.}

\item{color_palette}{Color palette to use for dendrogram. Defaults to
magma(1000, begin = 0.2).}

\item{xline}{Numeric value (or vector of numeric values) at which
to draw vertical line(s) along the dendrogram X-axis
(i.e., ANI values). Defaults to NULL (no vertical lines).}

\item{xlinecol}{Character (or vector of characters) corresponding to
ggplot2 color(s) for xline. Defaults to "#20A387FF".}

\item{xlinetype}{Character (or vector of characters) corresponding to
ggplot2 line type(s) for xline. Defaults to "dashed".}
}
\value{
medoid_genomes A data frame of medoid genomes detected at an ANI threshold (ANI_threshold).
If an ANI threshold of NULL supplied, this will be NULL.

cluster_assignments A data frame containing all genomes and the genomospecies clusters to which
they are assigned at an ANI threshold (ANI_threshold).
If an ANI threshold of NULL is supplied, this will be NULL.
}
\description{
This function constructs a dendrogram using pairwise ANI values
calculated between genomes and the dissimilarity metric described
by Carroll, Wiedmann, and Kovac (2019), yielding medoid genomes and
cluster assignments at an ANI threshold (if supplied by the user).
}
