% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetools.R
\name{phylo.discrete_trait_OTU}
\alias{phylo.discrete_trait_OTU}
\title{Highlight taxa in a phylogeny using a discrete variable}
\usage{
phylo.discrete_trait_OTU(phylo, trait_list = NULL,
  color_palette = viridis(n = length(trait_list)),
  phylo_layout = "circular", tip_label_size = 2, show_legend = T,
  legend_position = "left")
}
\arguments{
\item{phylo}{A phylo object}

\item{trait_list}{Named list, where names corresponds to traits and
vectors under each name correspond to taxa associated with that trait.
Defaults to NULL (no list is supplied).}

\item{color_palette}{Vector of trait color(s). Defaults to
one color per trait using the viridis color scale, i.e.,
viridis(n = length(trait_list)).}

\item{phylo_layout}{Any ggtree phylogeny layout.
Defaults to "circular".}

\item{tip_label_size}{Size of tip labels. Defaults to 2.}

\item{show_legend}{Display legend (True or False). Defaults to True.}

\item{legend_position}{Position of legend. Defaults to "left".}
}
\value{
A ggplot2 object
}
\description{
This function highlights taxa in a phylogeny using the groupOTU
 function in tidytree/ggtree.
}
